LPS-67908 Move
