+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-kaleo-runtime-api/src/main/java/com/liferay/portal/workflow/kaleo/runtime/activator/KaleoDefaults.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.runtime;
+package com.liferay.portal.workflow.kaleo.runtime.activator;
 
 /**
  * @author Inácio Nery
